main.py
import subprocess
import os
import threading
from huggingface_hub import hf_hub_download
from flask import Flask, request, jsonify
from github import Github
import torch

#main app code here

class Agent:
    def __init__(self, is_local):
        self.is_local = is_local
        self.memory = {}

def read_and_update_memory(self):
    """Read and update the agent's memory."""
    try:
        # Read memory from a file or database
        with open('memory.txt', 'r') as f:
            self.memory = json.load(f)

        # Update memory with new data
        self.memory['new_data'] = 'some_value'

        # Write memory back to file or database
        with open('memory.txt', 'w') as f:
            json.dump(self.memory, f)

    except Exception as e:
        print(f"Error reading or updating memory: {e}")


    def gather_system_logs(self):
        """Gather system logs for the agent."""
        # Implement the logic to gather system logs
        pass

    def greet_user(self):
        """Greet the user."""
        # Implement the logic to greet the user
        return "Hello, how can I assist you today?"

class App:
    def __init__(self):
        self.agents = []
        self.model = None
        self.app = Flask(__name__)
        self.github = Github()
        self.configure_routes()

    def configure_routes(self):
        """Configure the Flask routes."""
        @self.app.route('/chat', methods=['POST'])
        def chat():
            user_input = request.json.get('input')
            response = self.model_interaction(user_input)
            return jsonify({"response": response})

    def check_llama_installed(self):
        """Check if llama.cpp is installed."""
        try:
            result = subprocess.run(['llama.cpp', '--version'], capture_output=True, text=True)
            return result.returncode == 0
        except FileNotFoundError:
            return False

    def check_gpu_capabilities(self):
        """Check the GPU capabilities of the system."""
        if torch.cuda.is_available():
            gpu_name = torch.cuda.get_device_name(0)
            gpu_memory = torch.cuda.get_device_properties(0).total_memory / (1024 ** 3)  # in GB
            return {"gpu_name": gpu_name, "gpu_memory": gpu_memory}
        return None

    def install_llama(self):
        """Install llama.cpp."""
        try:
            subprocess.run(['git', 'clone', 'https://github.com/ggerganov/llama.cpp'], check=True)
            os.chdir('llama.cpp')
            subprocess.run(['make'], check=True)
            os.chdir('..')
            print("llama.cpp installed successfully.")
        except subprocess.CalledProcessError as e:
            print(f"Error during installation: {e}")

    def recommend_model(self, gpu_info):
        """Recommend a model based on GPU capabilities."""
        if gpu_info:
            if gpu_info['gpu_memory'] > 10:
                return "big-model"  # Placeholder name
            elif gpu_info['gpu_memory'] > 5:
                return "medium-model"
            else:
                return "small-model"
        return "small-model"

def download_model(self, model_name, directory="models"):
    """Download the specified model from HuggingFace."""
    try:
        os.makedirs(directory, exist_ok=True)
        model_path = hf_hub_download(repo_id=model_name, filename="pytorch_model.bin", cache_dir=directory)

        # Check if the downloaded file is a valid model file
        # This is a placeholder, you might want to use a more sophisticated method
        if not os.path.isfile(model_path):
            raise Exception("Downloaded file is not a valid model file")

        return model_path

    except Exception as e:
        print(f"Error downloading model: {e}")
        return None


    def run_model(self, model_path, gpu_info):
        """Run the model with appropriate settings."""
        try:
            command = ['llama.cpp', '--model', model_path]
            if gpu_info:
                command.extend(['--use-gpu', '--gpu-id', '0'])
            subprocess.run(command, check=True)
        except subprocess.CalledProcessError as e:
            print(f"Error running model: {e}")

    def model_interaction(self, user_input):
        """Interact with the model and return its response."""
        # Placeholder for actual model interaction
        return f"Model response to: {user_input}"

    def start_web_server(self):
        """Start the web server."""
        self.app.run(port=5000, debug=True)

    def create_agent(self, is_local):
        """Create a new agent."""
        agent = Agent(is_local)
        self.agents.append(agent)
        return agent

    def fork_repo(self, repo_name):
        """Fork a GitHub repository."""
        try:
            user = self.github.get_user()
            repo = self.github.get_repo(repo_name)
            user.create_fork(repo)
            print(f"Repository {repo_name} forked successfully.")
        except Exception as e:
            print(f"Error forking repository: {e}")

    def update_repo(self, repo_name, commit_message):
        """Update a GitHub repository with a new commit."""
        try:
            repo = self.github.get_repo(repo_name)
            # Implement the logic to update the repo
            # Placeholder for updating the repository
            print(f"Repository {repo_name} updated with message: {commit_message}")
        except Exception as e:
            print(f"Error updating repository: {e}")

    def main(self):
        """Main function to start the app."""
        llama_installed = self.check_llama_installed()
        gpu_info = self.check_gpu_capabilities()
        print(f"Llama.cpp installed: {llama_installed}")
        print(f"GPU Info: {gpu_info}")

        if not llama_installed:
            self.install_llama()

        model_name = self.recommend_model(gpu_info)
        model_path = self.download_model(model_name)
        if model_path:
            self.run_model(model_path, gpu_info)

        web_thread = threading.Thread(target=self.start_web_server)
        web_thread.start()
        print("Web chat interface running at http://localhost:5000")

if __name__ == "__main__":
    app = App()
    app.main()


def update_codebase():
    """Update the app's codebase to the latest version of the latest fork."""
    try:
        # Get the latest fork of the repository
        g = Github()
        repo = g.get_repo(f"username/{project_name}")
        fork = repo.get_forks()[0]

        # Clone the latest fork
        subprocess.run(['git', 'clone', fork.clone_url], check=True)

        # Replace the current codebase with the latest fork
        subprocess.run(['rm', '-rf', '*'], check=True)
        subprocess.run(['mv', f'{project_name}/*', '.'], check=True)
        subprocess.run(['rm', '-rf', project_name], check=True)

        print("Codebase updated successfully.")

    except Exception as e:
        print(f"Error updating codebase: {e}")

if __name__ == "__main__":
    # Add your main function here
    pass
